Rename SigV4AwsProperties and change the package
